requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
cloudinary==1.8.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.1
Django==1.11.1
django-ckeditor==5.3.0
django-cors-headers==2.2.0
django-js-asset==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.6.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
olefile==0.45.1
openapi-codec==1.3.2
pbr==3.1.1
Pillow==4.2.0
psycopg2-binary==2.7.4
PyJWT==1.6.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
whitenoise==3.2
